Made bar chart interactive, restructured files
