+++ root/index.html
@@ -48,7 +48,13 @@
     <!-- scripts -->
     <script src="js/d3init.js"></script>
     <script src="js/script.js"></script>
-    <script src="js/comparison_card.js"></script>
+    <script src="js/Comparison_Card_Components/select_search_component.js" type="module"></script>
+    <script src="js/Comparison_Card_Components/radar_chart_component.js" type="module"></script>
+    <script src="js/Comparison_Card_Components/info_card_component.js" type="module"></script>
+    <script src="js/Comparison_Card_Components/popularity_score_component.js" type="module"></script>
+    <script src="js/Comparison_Card_Components/bar_chart_component.js" type="module"></script>
+    <script src="js/Comparison_Card_Components/comparison_card_component.js" type="module"></script>
+    <script src="js/comparison_card.js" type="module"></script>
     <script src="js/GenreComparisonChart.js"></script>
 </body>
